functions/requirements.txt
google-cloud-firestore==1.6.2
Flask==1.0.2
firebase_admin==4.0.1